requirements.txt
pandas==2.1.4
gitpython==3.1.40
alive-progress==3.1.5
datasets==2.14.7
transformers==4.36.2
tokenizers==0.15.0
matplotlib==3.8.2
seaborn==0.13.0
pycparser==2.21
tqdm==4.66.1
jsonlines==3.1.0
colorama==0.4.6
openai==2.8.1
